Many changes have been made. There has been a bug that the !newteam command doesn't work with no members passed. There must have either been a valid or invalid discord member, otherwise the command would fail. This is no longer an issue.
The !addplayers command is now a valid command. This adds one or more players to the team in the command. I extracted as much overlapping behavior as I could from the !newteam command. This largely consists of checking the validity of the names provided in the commands, and adding valid players to the team.

This does not add the associated role, but the plan is that it will in the next commit. In the meantime, roles are not created when a !newteam command is called. This will be corrected when the roles are managed properly with the !addplayers command.